Add tests to GiftCardActivate and GiftCardDeactivate mutations
